updated hmm coloring
It was previously assumed you would use the scheme of "X..." to name a state, where the X is a character followed by an integer. For example, U2 or U21 may indicate undersegmentation, M5 and M105 may indicate matches, and so on. Now, you don't need to follow that explicitly, but your coloring will be by mapping states to colors by their hash value, not their order, which may not be desired.