segmenter, file plotting, core duration
(1) The segmenter has been changed to handle arbitrary filtration, and
reparameterized to use false positives per second as opposed to
oversegmentation rate

(2) File plotting now lets you plot regions of a file, with or without
event detection run, with different arguments for the events and
non-event regions.

(3) Previously you the number of samples per second was hard coded in
the duration measurements. Now it is taken from the file, and duration
is never explicitly calculated.